refactor: cr proposed changes
